+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Atlas Content Modeler Changelog
 
+## Unreleased
+### Changed
+- Text fields can now use “title” as their API identifier if “use this field as the entry title” is ticked.
+
 ## 0.17.0 - 2022-05-05
 ### Added
 - Mutations support to create, update and delete entries via GraphQL. All fields are supported except media and relationships for now. Find examples at https://github.com/wpengine/atlas-content-modeler/blob/main/docs/mutations/index.md.